+++ services/preview/storage/mgmt/2018-07-01-preview/storage/storageapi/interfaces.go
@@ -19,7 +19,7 @@ package storageapi
 
 import (
 	"context"
-	"github.com/Azure/azure-sdk-for-go/services/storage/mgmt/2018-07-01/storage"
+	"github.com/Azure/azure-sdk-for-go/services/preview/storage/mgmt/2018-07-01-preview/storage"
 	"github.com/Azure/go-autorest/autorest"
 )